Commit App backend du 27 Juin

+++ src/main/java/com/macsecurite/macsecurite/model/HistoriqueIntervention.java
@@ -19,7 +19,7 @@ public class HistoriqueIntervention {
      */
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long idHistorique;
+    private Integer idHistorique;
 
     /**
      * Intervention associée à cet historique.
