+++ src/main/java/com/macsecurite/macsecurite/repository/CategorieProduitRepository.java
@@ -7,6 +7,5 @@ import org.springframework.stereotype.Repository;
  * Interface de repository responsable de l'accès aux données pour l'entité CategorieProduit dans la base de données.
  */
 @Repository
-public interface CategorieProduitRepository extends JpaRepository<CategorieProduit, Long> {
-    // Ajout des méthodes supplémentaires si nécessaire
+public interface CategorieProduitRepository extends JpaRepository<CategorieProduit, Integer> {
 }
