+++ src/main/java/com/macsecurite/macsecurite/repository/CommandeRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
  * Interface de repository responsable de l'accès aux données pour l'entité Commande dans la base de données.
  */
 @Repository
-public interface CommandeRepository extends JpaRepository<Commande, Long> {
+public interface CommandeRepository extends JpaRepository<Commande, Integer> {
     // Ajout des méthodes supplémentaires si nécessaire
 }
