+++ src/main/java/com/macsecurite/macsecurite/repository/HistoriqueInterventionRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
  * Interface de repository responsable de l'accès aux données pour l'entité HistoriqueIntervention dans la base de données.
  */
 @Repository
-public interface HistoriqueInterventionRepository extends JpaRepository<HistoriqueIntervention, Long> {
+public interface HistoriqueInterventionRepository extends JpaRepository<HistoriqueIntervention, Integer> {
     // Ajout des méthodes supplémentaires si nécessaire
 }
