+++ src/main/java/com/macsecurite/macsecurite/repository/TechnicienRepository.java
@@ -10,6 +10,6 @@ import org.springframework.stereotype.Repository;
  */
 @Repository
 // Ce repository est responsable de l'accès aux données pour l'entité Technicien dans la base de données.
-public interface TechnicienRepository extends JpaRepository<Technicien, Long> {
+public interface TechnicienRepository extends JpaRepository<Technicien, Integer> {
     // Ajout des méthodes supplémentaires si nécessaire
 }
